build: Install xml files into wf-shell subdirectory
The wf-shell xml files were installed to the same directory as the
wayfire xml files, which was causing wayfire and wf-shell sections
to be written to both config files when save_configuration was called.
This puts the xml files in a subdirectory of the metadata directory,
so the correct sections are written to the respective config files.