window-list: use DesktopAppInfo to find apps for icons
The icons in the window list are loaded by parsing the .desktop
files for a few variations of the window's app ID in a few known
locations. This doesn't look everywhere it should - for example,
it doesn't look into ~/.local/share/applications or look at
XDG_DATA_DIRS.

Use the standard Gio::DesktopAppInfo::create instead.