window-list: Fix minimize targets after closing a toplevel
Before, closing a toplevel and unminimizing another would unminimize from
the target position before closing the other toplevel. Sending the rectangle
hints after removing the button from the Gtk::Box did not work. Instead, we
send the rectangle hints from the layout vfunc allocation. This required
a lot of refactoring, but now the code is simpler and easier to work with.