panel : add brightness control widget
Handles brightness for displays exposed in sysfs, and external monitors that
implement MCCS. MCCS control depends on libddcutil, and is not built if it is
not detected.
Presents a slider per supported device in a popover on click. If we can detect
which display a device corresponds to, we isolate them in the view of the panel
of this display, and scrolling on the widget changes it.
Has the classic niceties of scrolling and pop out on detected change, although
we don’t have a way to watch for changes on MCCS displays.

shell: added signal_monitor_list_changed for post-change callback
shell: reinit hotspots only on post-change monitor list, removing some race condition crashes

Co-authored-by: trigg <triggerhapp@gmail.com>